2-strlen.c

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -5,9 +5,9 @@
  * @str: The string
  * Return: The lenght of @str
  */
-size_t _strlen(const char *str)
+int _strlen(const char *str)
 {
-	size_t len = 0;
+	int len = 0;
 
 	while (*str++)
 	{
